% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestdist_gcms.R
\name{closestdist_gcms}
\alias{closestdist_gcms}
\title{Distance between GCMs}
\usage{
closestdist_gcms(
  s,
  var_names = c("bio_1", "bio_12"),
  study_area = NULL,
  scale = TRUE,
  k = NULL,
  method = "euclidean",
  minimize_difference = TRUE,
  max_difference = NULL
)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. A vector with names of the bioclimatic variables to compare OR 'all'.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}

\item{scale}{Boolean. Apply center and scale in data? Default is TRUE.}

\item{k}{Number of GCMs. If NULL (standard), stopping criteria are applied.}

\item{method}{The distance method to use. Default is "euclidean". Possible values are "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman" or "kendall". See ?dist_gcms.}

\item{minimize_difference}{Boolean. If k = NULL, function will search for best value of k. Standard is TRUE.}

\item{max_difference}{Numeric. Distance threshold to stop searching. Standard is NULL.}
}
\value{
Set of GCMs that have mean distance closer to the mean of all GCMs provided in s.
}
\description{
This function compares future climate projections from multiple Global Circulation Models (GCMs) based on their similarity in terms of variables.
}
\details{
\code{minimize_difference} will search for the optimum value of k by incorporating new GCMs
to the subset until the mean distance of the subset starts to move away from the mean distance from
all GCMs (global distance).
By setting a \code{max_difference} value, the function will test if the mean distance between GCMs
in the subset is lower than max_difference. If TRUE, the function returns the given subset, otherwise
it will keep searching for better results.
}
\examples{
\dontrun{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")

closestdist_gcms(s, method = "euclidean")
}

}
\seealso{
\code{\link{cor_gcms}} \code{\link{dist_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
