% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gcms.R
\name{cor_gcms}
\alias{cor_gcms}
\title{Compute and plot correlation matrix for a set of General Circulation Models.}
\usage{
cor_gcms(x, method = "pearson")
}
\arguments{
\item{x}{A raster stack of General Circulation Models.}

\item{method}{The correlation method to use. Default is "pearson". Possible values are "pearson", "kendall" and "spearman".}
}
\value{
A correlation matrix plot.
}
\description{
Compute and plot correlation matrix for a set of General Circulation Models.
}
\examples{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
s <- transform_gcms(s, var_names, study_area)
flattened_gcms <- flatten_gcms(s)
cor_gcms(quick_example, method = "spearman")

}
\seealso{
\code{\link{transform_gcms}} \code{\link{flatten_gcms}} \code{\link{summary_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
