% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Distance Matrix}
\usage{
distance_matrix(
  s,
  var_names = c("bio_1", "bio_12"),
  study_area = NULL,
  method = "euclidean"
)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. The names of the bioclimatic variables to compare.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}

\item{method}{The correlation method to use. Default is "pearson". Possible values are "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman" or "kendall".}
}
\value{
A list with two items: distances (the distances between GCMs) and heatmap (a plot).
}
\description{
This function compares future climate projections from multiple Global Circulation Models (GCMs) based on their similarity in terms of bioclimatic variables. The function calculates distance metrics and plot it on a heatmap.
}
\examples{
# compare GCMS
compare_gcms(folder_future_rasters_gcms = "path/to/folder",
study_area = raster("path/to/raster"),
var_names = c('bio_1', 'bio_12'),
gcm_names = c('gcm1', 'gcm2', 'gcm3'),
k = 3)

}
