% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_gcms.R
\name{env_gcms}
\alias{env_gcms}
\title{General Circulation Model (GCM) environmental distribution}
\usage{
env_gcms(
  s,
  var_names = c("bio_1", "bio_12"),
  study_area = NULL,
  highlight = "sum",
  resolution = 25
)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. A vector with names of the bioclimatic variables to compare OR 'all'.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}

\item{highlight}{Character. A vector with names of gcms to be highlighted. In this case, the sum of all but chosen GCMs will appear in grey.}

\item{resolution}{Res}
}
\value{
a data frame with the summary statistics for each variable
}
\description{
This function plots GCMs data in environmental space, possibly highlighting clusters or specific GCMs.
}
\examples{
\dontrun{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
env_gcms(s, var_names, study_area, highlight = "sum")
env_gcms(s, var_names, study_area, highlight = c("cr", "ml", "uk"))
}

}
\seealso{
\code{\link{summary_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
