% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_gcms.R
\name{flatten_gcms}
\alias{flatten_gcms}
\title{Flatten GCMs}
\usage{
flatten_gcms(s)
}
\arguments{
\item{s}{A list of transformed data frames representing GCMs.}
}
\value{
A named list of flattened \code{data.frame} objects.
}
\description{
Scale and flatten a list of General Circulation Models (GCMs) from rasters to vectors.
}
\examples{
\dontrun{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
s <- transform_gcms(s, var_names, study_area)
flattened_gcms <- flatten_gcms(s)
}

}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
