% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcms_data.R
\docType{data}
\name{gcms_data}
\alias{gcms_data}
\title{General Circulation Models (GCMs) data}
\format{
## `gcms_data`
A \code{list} of \code{SpatRaster}:
\describe{
  \item{names}{Bioclimatic variables}
  \item{resolution}{10 arc-minutes}
  \item{EPSG}{4326}
}
}
\source{
<https://www.worldclim.org/>
}
\usage{
gcms_data
}
\description{
A \code{list} object with 23 GCMs projections to the SSP5-8.5/2090 scenario obtained from WorldClim 2.1.
}
\keyword{datasets}
