% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_gcms.R
\name{hclust_gcms}
\alias{hclust_gcms}
\title{Hierarchical Clustering of GCMs}
\usage{
hclust_gcms(x, k = 3, n = 1000)
}
\arguments{
\item{x}{A list of stacks containing GCMs.}

\item{k}{The number of clusters to identify.}

\item{n}{The number of values to use in the clustering (default: 1000).}
}
\value{
A dendrogram visualizing the clusters.
}
\description{
Given a list of stacks containing GCMs, this function performs hierarchical clustering on a random subset of the raster values and produces a dendrogram visualization of the clusters.
}
\examples{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
s <- transform_gcms(s, var_names, study_area)
flattened_gcms <- flatten_gcms(s)
dend <- hclust_gcms(stack, k=4, n=500)
plot(dend)

}
\seealso{
\code{\link{transform_gcms}} \code{\link{flatten_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
