% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gcms.R
\name{import_gcms}
\alias{import_gcms}
\title{Import GCM data to R}
\usage{
import_gcms(
  path = "input_data/WorldClim_data_gcms",
  extension = ".tif",
  recursive = TRUE,
  gcm_names = NULL,
  var_names = NULL
)
}
\arguments{
\item{path}{A string with the path to GCM files.}

\item{extension}{Extension of stack files. Standard is ".tif", which is the extension from WorldClim 2.1.}

\item{recursive}{Logical. Should the function import stacks recursively (i.e. search files from folders within folders)? Standard is TRUE.}

\item{gcm_names}{A vector with names to be addressed to each GCM.}

\item{var_names}{A vector with names to be addressed to each variable.}
}
\value{
A list of stacks, with each element of the list corresponding to a GCM from given path.
}
\description{
This function imports GCM stack files from folder to R.
}
\examples{
\dontrun{
s <- import_gcms(
  path = "input_data/WorldClim_data_future", extension = ".tif",
  recursive = TRUE, gcm_names = NULL
)
}

}
\seealso{
\code{\link{worldclim_data}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
