% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_gcms.R
\name{kmeans_gcms}
\alias{kmeans_gcms}
\title{Perform k-means clustering on GCMs}
\usage{
kmeans_gcms(
  s,
  var_names = c("bio_1", "bio_12"),
  study_area = NULL,
  k = 3,
  method = NULL
)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. The names of the bioclimatic variables to compare.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}

\item{k}{Number of clusters.}

\item{method}{The method for distance matrix computation. Standard value is "euclidean". Possible values are: "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". If NULL, will perform the clustering on raw variables data.}
}
\value{
A scatter plot of the resulting clusters.
}
\description{
This function performs k-means clustering on a distance matrix and produces a scatter plot of the resulting clusters.
}
\examples{

s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")

kmeans_gcms(s, k=3)

}
\seealso{
\code{\link{transform_gcms}} \code{\link{flatten_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
