% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo_gcms.R
\name{montecarlo_gcms}
\alias{montecarlo_gcms}
\title{Perform Monte Carlo permutations on GCMs}
\usage{
montecarlo_gcms(
  s,
  var_names = c("bio_1", "bio_12"),
  study_area = NULL,
  scale = TRUE,
  perm = 10000,
  dist_method = "euclidean",
  clustering_method = "closestdist",
  ...
)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. A vector with names of the bioclimatic variables to compare OR 'all'.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}

\item{scale}{Boolean. Apply center and scale in data? Default is TRUE.}

\item{perm}{Number of permutations.}

\item{dist_method}{The method for distance matrix computation. Standard value is "euclidean". Possible values are: "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". If NULL, will perform the clustering on raw variables data.}

\item{clustering_method}{The method for clustering. Standard value is "closestdist". Possible values are: "kmeans", "hclust" or "closestdist".}

\item{...}{Arguments to pass to the clustering function.}
}
\value{
A violin plot of the result. Dashed red line and red dots represent the mean absolute distance between subset of GCMs using the clustering approach. Violin plot is built with Monte Carlo permutations, selecting random subsets of GCMs from the given set.
}
\description{
This function performs Monte Carlo permutations on a distance matrix and produces a violin plot of the resulting mean distance between subsets of the distance matrix.
}
\examples{

\dontrun{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
montecarlo_gcms(s, var_names, study_area)
}

}
\seealso{
\code{\link{hclust_gcms}} \code{\link{env_gcms}} \code{\link{kmeans_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
