% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldClim_data.R
\name{WorldClim_data}
\alias{WorldClim_data}
\title{Download WorldClim v.2.1 bioclimatic data}
\usage{
WorldClim_data(period = 'current', variable = 'bioc', year = '2030', gcm = 'mi', ssp = '126', resolution = 10)
}
\arguments{
\item{period}{Can be 'current' or 'future'.}

\item{variable}{Allows to specify which variables you want to retrieve Possible entries are:
'tmax',''tmin,'prec' and/or 'bioc'.}

\item{year}{Specify the year you want to retrieve data. Possible entries are:
'2030', '2050', '2070' and/or '2090'. You can  use a vector to provide more than one entry.}

\item{gcm}{GCMs to be considered in future scenarios. You can use a vector to provide more than one entry.
| **CODE** | **GCM** |
| ---- | ---------------- |
| ac  | ACCESS-CM2 |
| ae  | ACCESS-ESM1-5 |
| bc  | BCC-CSM2-MR |
| ca  | CanESM5 |
| cc  | CanESM5-CanOE |
| ce  | CMCC-ESM2 |
| cn  | CNRM-CM6-1 |
| ch  | CNRM-CM6-1-HR |
| cr  | CNRM-ESM2-1 |
| ec  | EC-Earth3-Veg |
| ev  | EC-Earth3-Veg-LR |
| fi  | FIO-ESM-2-0 |
| gf  | GFDL-ESM4 |
| gg  | GISS-E2-1-G |
| gh  | GISS-E2-1-H |
| hg  | HadGEM3-GC31-LL |
| in  | INM-CM4-8 |
| ic  | INM-CM5-0 |
| ip  | IPSL-CM6A-LR |
| me  | MIROC-ES2L |
| mi  | MIROC6 |
| mp  | MPI-ESM1-2-HR |
| ml  | MPI-ESM1-2-LR |
| mr  | MRI-ESM2-0 |
| uk  | UKESM1-0-LL |}

\item{ssp}{SSPs for future data. Possible entries are: '126', '245', '370' and/or '585'.
You can use a vector to provide more than one entry.}

\item{resolution}{You can select one resolution from the following alternatives: 10, 5, 2.5 OR 30.}
}
\description{
This function allows to download data from WorldClim v.2.1 (https://www.worldclim.org/data/index.html) considering multiple GCMs, time periods and SSPs.
}
\details{
This function will create a folder entitled 'WorldClim_data'. All the data downloaded will be stored in this folder. Note that, despite being possible to retrieve a lot of data at once, it is not recommended to do so, since the data is very heavy.
}
\examples{
\dontrun{# download data from multiple periods:
year <- c(2050, 2090)
WorldClim_data('bioc',year, 'mi', 126, 10)

# download data from one specific period
WorldClim_data('bioc',2070,'mi',585,10)}

}
\references{
https://www.worldclim.org/data/index.html
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
