% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestdist_gcms.R
\name{closestdist_gcms}
\alias{closestdist_gcms}
\title{Distance between GCMs}
\usage{
closestdist_gcms(
  s,
  var_names = c("bio_1", "bio_12"),
  study_area = NULL,
  method = "euclidean"
)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. A vector with names of the bioclimatic variables to compare OR 'all'.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}

\item{method}{The correlation method to use. Default is "euclidean". Possible values are "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman" or "kendall".}
}
\value{
Set of two GCMs that have mean distance closer to the mean of all GCMs provided in s.
}
\description{
This function compares future climate projections from multiple Global Circulation Models (GCMs) based on their similarity in terms of variables.
}
\examples{
\dontrun{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")

closestdist_gcms(s, method = "euclidean")
}

}
\seealso{
\code{\link{transform_gcms}} \code{\link{flatten_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
