% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gcms.R
\name{compare_gcms}
\alias{compare_gcms}
\title{Compare GCMS}
\usage{
compare_gcms(
  s,
  var_names = c("bio_1", "bio_12"),
  study_area = NULL,
  k = 3,
  scale = TRUE,
  clustering_method = "closestdist"
)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. A vector with names of the bioclimatic variables to compare OR 'all'.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}

\item{k}{Numeric. The number of clusters to use for k-means clustering.}

\item{scale}{Boolean. Apply center and scale in data? Default is TRUE.}
}
\value{
A list with two items: suggested_gcms (the names of the GCMs suggested for further analysis) and statistics_gcms (a grid of plots).
}
\description{
This function compares future climate projections from multiple Global Circulation Models (GCMs) based on their similarity in terms of bioclimatic variables. The function clusters the GCMs using k-means clustering and hierarchical clustering, calculates the Euclidean distance matrix, and generates plots for the clusters and the distance matrix.
}
\examples{
\dontrun{
# compare GCMS
compare_gcms(
  s = import_gcms("path/to/folder"),
  study_area = terra::vect("path/to/gpkg"),
  var_names = c("bio_1", "bio_12"),
  gcm_names = c("gcm1", "gcm2", "gcm3"),
  k = 3
)
}

}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
