% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gcms.R
\name{cor_gcms}
\alias{cor_gcms}
\title{Compute and plot correlation matrix for a set of General Circulation Models.}
\usage{
cor_gcms(
  s,
  var_names = c("bio_1", "bio_12"),
  study_area = NULL,
  scale = TRUE,
  method = "pearson"
)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. A vector with names of the bioclimatic variables to compare OR 'all'.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}

\item{scale}{Boolean. Apply center and scale in data? Default is TRUE.}

\item{method}{The correlation method to use. Default is "pearson". Possible values are "pearson", "kendall" or "spearman".}
}
\value{
A list with two items: cor_matrix (the correlations between GCMs) and cor_plot (a correlation plot).
}
\description{
Compute and plot correlation matrix for a set of General Circulation Models.
}
\examples{
\dontrun{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
cor_gcms(s, var_names, study_area, method = "spearman")
}

}
\seealso{
\code{\link{transform_gcms}} \code{\link{flatten_gcms}} \code{\link{summary_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
