% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gcms.R
\name{dist_gcms}
\alias{dist_gcms}
\title{Distance between GCMs}
\usage{
dist_gcms(
  s,
  var_names = c("bio_1", "bio_12"),
  study_area = NULL,
  method = "euclidean"
)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. The names of the bioclimatic variables to compare.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}

\item{method}{The correlation method to use. Default is "pearson". Possible values are "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman" or "kendall".}
}
\value{
A list with two items: distances (the distances between GCMs) and heatmap (a plot).
}
\description{
This function compares future climate projections from multiple Global Circulation Models (GCMs) based on their similarity in terms of bioclimatic variables. The function calculates distance metrics and plot it on a heatmap.
}
\examples{

s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")

dist_gcms(s, method='euclidean')

}
\seealso{
\code{\link{transform_gcms}} \code{\link{flatten_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
