% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_gcms.R
\name{hclust_gcms}
\alias{hclust_gcms}
\title{Hierarchical Clustering of GCMs}
\usage{
hclust_gcms(
  s,
  var_names = c("bio_1", "bio_12"),
  study_area = NULL,
  k = 3,
  n = NULL
)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. A vector with names of the bioclimatic variables to compare OR 'all'.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}

\item{k}{The number of clusters to identify.}

\item{n}{The number of values to use in the clustering. If NULL (default) all data is used.}
}
\value{
A dendrogram visualizing the clusters.
}
\description{
This function performs hierarchical clustering on a random subset of the raster values and produces a dendrogram visualization of the clusters.
}
\examples{

s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
dend <- hclust_gcms(stack, k=4, n=500)
plot(dend)

}
\seealso{
\code{\link{transform_gcms}} \code{\link{flatten_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
