% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_gcms.R
\name{kmeans_gcms}
\alias{kmeans_gcms}
\title{Perform k-means clustering on GCMs}
\usage{
kmeans_gcms(dist_matrix, k = 3, method = "euclidean")
}
\arguments{
\item{dist_matrix}{A list of stacks representing GCMs.}

\item{k}{Number of clusters.}

\item{method}{The method for distance matrix computation. Standard value is "euclidean". Possible values are: "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}
}
\value{
A scatter plot of the resulting clusters.
}
\description{
This function performs k-means clustering on a distance matrix and produces a scatter plot of the resulting clusters.
}
\examples{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
s <- transform_gcms(s, var_names, study_area)
kmeans_gcms(s, k=3)

}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
