% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_gcms.R
\name{summary_gcms}
\alias{summary_gcms}
\title{Summarize General Circulation Model (GCM) data}
\usage{
summary_gcms(s)
}
\arguments{
\item{s}{a transformed list of stacks representing GCM data}
}
\value{
a data frame with the summary statistics for each variable
}
\description{
This function summarizes GCM data by calculating several statistics for each variable
}
\examples{
s <- list(stack("file1.nc"), stack("file2.nc"))
summary_gcms(s)

s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
s <- transform_gcms(s, var_names, study_area)
summary_gcms(s)

}
\seealso{
\code{\link{transform_gcms}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
\keyword{GCM,}
\keyword{raster}
\keyword{summary,}
