% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_gcms.R
\name{transform_gcms}
\alias{transform_gcms}
\title{Transform General Circulation Model (GCM) stacks}
\usage{
transform_gcms(s, var_names = c("bio_1", "bio_12"), study_area = NULL)
}
\arguments{
\item{s}{A list of stacks of General Circulation Models.}

\item{var_names}{Character. The names of the bioclimatic variables to compare.}

\item{study_area}{Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.}
}
\value{
A list of data frames, with each element of the list corresponding to a GCM in the input list.
}
\description{
This function transforms a list of stacks of GCMs by subsetting it to only include the variable names specified in \code{var_names}, reprojecting it to match the CRS of \code{study_area}, cropping and masking it to \code{study_area}, and returning a list of data frames.
}
\examples{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
t <- transform_gcms(s, var_names, study_area)

}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
